factor into utility function
