+++ context/app/static/js/components/workspaces/JobDetails.jsx
@@ -1,43 +1,18 @@
 import React from 'react';
 
 import { LightBlueLink } from 'js/shared-styles/Links';
-
-const ACTIVE = 'Active';
-const ACTIVATING = 'Activating';
-const INACTIVE = 'Inactive';
-
-function getDisplayStatus(status) {
-  return (
-    {
-      pending: ACTIVATING,
-      running: ACTIVE,
-    }[status] || INACTIVE
-  );
-}
-
-function getJobUrl(job) {
-  const { url_domain, url_path } = job.job_details.current_job_details.connection_details;
-  return `${url_domain}${url_path}`;
-}
+import { condenseJobs } from './utils';
 
 function JobDetails({ jobs }) {
-  const displayJobs = jobs.map((job) => {
-    const diplayJob = { ...job };
-    diplayJob.status = getDisplayStatus(job.status);
-    return diplayJob;
-  });
+  const job = condenseJobs(jobs);
 
-  const bestJob = [ACTIVE, ACTIVATING, INACTIVE]
-    .map((status) => displayJobs.find((job) => job.status === status))
-    .find((job) => job);
-
-  if (!bestJob) {
+  if (!job) {
     return null;
   }
-  if (bestJob.status === ACTIVE) {
-    return <LightBlueLink href={getJobUrl(bestJob)}>Status: {ACTIVE}</LightBlueLink>;
+  if (job.url) {
+    return <LightBlueLink href={job.url}>Status: {job.status}</LightBlueLink>;
   }
-  return `Status: ${bestJob.status}`;
+  return `Status: ${job.status}`;
 }
 
 export default JobDetails;
