+++ context/app/static/js/components/workspaces/utils.js
@@ -59,4 +59,42 @@ function mergeJobsIntoWorkspaces(jobs, workspaces) {
   return workspaces;
 }
 
-export { createNotebookWorkspace, startJob, mergeJobsIntoWorkspaces };
+function condenseJobs(jobs) {
+  const ACTIVE = 'Active';
+  const ACTIVATING = 'Activating';
+  const INACTIVE = 'Inactive';
+
+  function getDisplayStatus(status) {
+    return (
+      {
+        pending: ACTIVATING,
+        running: ACTIVE,
+      }[status] || INACTIVE
+    );
+  }
+
+  function getJobUrl(job) {
+    const { url_domain, url_path } = job.job_details.current_job_details.connection_details;
+    return `${url_domain}${url_path}`;
+  }
+
+  const displayJobs = jobs.map((job) => {
+    const diplayJob = { ...job };
+    diplayJob.status = getDisplayStatus(job.status);
+    return diplayJob;
+  });
+
+  const bestJob = [ACTIVE, ACTIVATING, INACTIVE]
+    .map((status) => displayJobs.find((job) => job.status === status))
+    .find((job) => job);
+
+  if (!bestJob) {
+    return null;
+  }
+  if (bestJob.status === ACTIVE) {
+    return { status: ACTIVE, url: getJobUrl(bestJob) };
+  }
+  return { status: bestJob.status };
+}
+
+export { createNotebookWorkspace, startJob, mergeJobsIntoWorkspaces, condenseJobs };
